exponent môže byť len celé číslo

+++ zadanie01.c
@@ -5,7 +5,7 @@
 #define OK 1
 #define FAIL 0
 
-int mocnina(int mocnenec, int exponent);
+double mocnina(double mocnenec, int exponent);
 
 char mta_odmocnina(int m, double x, double *res) // mta odmocnina z x
 {
@@ -33,7 +33,7 @@ char mta_odmocnina(int m, double x, double *res) // mta odmocnina z x
      return OK;
 }
 
-int mocnina(int mocnenec, int exponent)
+double mocnina(double mocnenec, int exponent)
 {
 
      if (exponent == 0 || mocnenec == 1.0)
